pyproject.toml
[tool.poetry]
name = "legal-clause-finder"
version = "0.1.0"
description = ""
authors = ["Abhishek Mahajan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic = "^2.10.3"
uvicorn = "^0.33.0"
python-multipart = "^0.0.19"
pdf2image = "^1.17.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
